Add EncryptedStreamCodec to allow for encrypting tuples
The EncryptedStreamCodec allows to encrypt all the tuples before sending
them downstream.
It is intended to be used as a decorator on existing streamcodes, such
as the KryoSerializableStreamCodec. An example usage could look like the
following:

`StreamCodec<String> streamCodec =
   new EncryptedStreamCodec<>(new KryoSerializableStreamCodec<String>());`

The current mechanism uses AES for encrypting all the tuples. The secret
key gets serialized and sent to the nodes where it is needed with the
rest of the streamcodec, which means this whole mechanism is save only
when the corresponding transport security for hadoop rpc is in place.

This commit additionally adds some tests to check if the encryption /
decryption works as expected